backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
httpx>=0.25.0
websockets>=12.0
python-multipart>=0.0.6
redis>=5.0.0
aioredis>=2.0.1
sse-starlette>=2.0.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.0
pinecone-client>=2.2.0
langchain-pinecone>=0.0.1
langchain-chroma>=0.0.1

# Memory and ML
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Workflow and orchestration
langgraph>=0.0.20
networkx>=3.0

# Parsing and chunking
beautifulsoup4>=4.12.0
pypdf>=3.0.0
markdown>=3.5.0
tiktoken>=0.5.0

# Monitoring and evaluation
prometheus-client>=0.19.0
tenacity>=8.2.0